% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwichMethods.R
\name{bread.singleR}
\alias{bread.singleR}
\title{Extracting bread matrix for singleR class}
\usage{
\method{bread}{singleR}(object, ...)
}
\arguments{
\item{object}{an object representing a fitted model.}

\item{...}{additional optional arguments passed to the following functions:
\itemize{
\item \code{stats::vcov} -- for extracting "the usual" variance-covariance matrix, \code{vcov.singleR} has one additional argument \code{type} with values \code{"Fisher"} \code{"observedInform"}, defaults to the one specified in \code{control.pop.var} specified in call for object.
}}
}
\value{
A bread matrix, i.e. a hessian based estimation of variance-covariance matrix scaled by degrees of freedom.
}
\description{
Extracting bread matrix for singleR class
}
\examples{
Model <- estimate_popsize(
formula = capture ~ ., 
data = netherlandsimmigrant, 
model = ztpoisson, 
method = "IRLS")
sandwich::bread(Model)
vcov(Model)
# This function just scales.
all(vcov(Model) * nrow(netherlandsimmigrant) == sandwich::bread(Model))
# We can choose Fisher information matrix instead of default observed information matrix.
vcov(Model, "Fisher")
sandwich::bread(Model, type = "Fisher")
all(vcov(Model, "Fisher") * nrow(netherlandsimmigrant) == sandwich::bread(Model, type = "Fisher"))
}
\seealso{
\code{\link[sandwich:bread]{sandwich::bread()}} \code{\link[=control.pop.var]{control.pop.var()}}
}
\author{
Piotr Chlebicki, Maciej Beręsewicz

An S3class for \code{sandwich::bread} to handle \code{singleR} objects. This function was developed based on \code{sandwich:::bread.glm}
}
