% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chao.R
\name{chao}
\alias{chao}
\title{Chao model for population estimate}
\usage{
chao()
}
\value{
A object of class "family" containing objects \cr
make_minusloglike(y,X) - for creating negative likelihood function \cr
make_gradient(y,X) - for creating gradient function \cr
make_hessian(X) - for creating hessian \cr
linkfun - a link function to connect between linear predictor and model parameter in regression and a name of link function\cr
linkinv - an inverse function of link \cr
dlink - a 1st derivative of link function \cr
mu.eta,variance - Expected Value and variance \cr
aic - for aic computation\cr
valedmu, valideta - for checking if regression arguments and valid\cr
family - family name\cr
Where: \cr
y is a vector of observed values \cr
X is a matrix / data frame of covariates
}
\description{
Chao model for population estimate
}
