% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{controlMethod}
\alias{controlMethod}
\title{Control parameters for regression}
\usage{
controlMethod(
  epsilon = 1e-08,
  maxiter = 1000,
  verbose = 0,
  printEveryN = 1,
  start = NULL,
  alphaStart = NULL,
  omegaStart = NULL,
  piStart = NULL,
  optimMethod = "L-BFGS-B",
  silent = FALSE,
  optimPass = FALSE,
  stepsize = 1,
  checkDiagWeights = TRUE,
  weightsEpsilon = .Machine$double.eps^0.75,
  momentumFactor = 0,
  saveIRLSlogs = FALSE,
  useZtpoissonAsStart = FALSE,
  momentumActivation = 5,
  criterion = c("coef", "abstol", "reltol")
)
}
\arguments{
\item{epsilon}{Tolerance for fitting algorithms by default \code{1e-8}.}

\item{maxiter}{Maximum number of iterations.}

\item{verbose}{Value indicating whether to trace steps of fitting algorithm for
\code{IRLS} fitting method different values of verbose give the following information:
\itemize{
\item 1 -- Returns information on the number of current iteration and current log-likelihood.
\item 2 -- Returns information on vector of regression parameters at current iteration (and all of the above).
\item 3 -- Returns information on reduction of log-likelihood at current iteration (and all of the above).
\item 4 -- Returns information on value of log-likelihood function gradient at current iteration (and all of the above).
\item 5 -- Returns information on convergence criterion and values that are taken into account when considering convergence (and all of the above).
}
if \code{optim} method was chosen verbose will be passed to \code{\link[stats:optim]{stats::optim()}} as trace.}

\item{printEveryN}{Integer value indicating how often to print information
specified in \code{verbose}, by default set to \code{1}.}

\item{start}{initial parameters for regression associated with main formula
specified in function call if \code{NULL} they will be derived from simple poisson regression.}

\item{alphaStart}{initial parameters for dispersion parameter if applies.}

\item{omegaStart}{initial parameters for inflation parameter if applies.}

\item{piStart}{initial parameters for probability parameter if applies.}

\item{optimMethod}{method of \code{\link[stats:optim]{stats::optim()}} used L-BFGS-B is the default
except for negative binomial and one inflated models where \code{"Nelder-Mead"} is used.}

\item{silent}{Logical, indicating whether warnings in \code{IRLS} method should be suppressed.}

\item{optimPass}{Optional list of parameters passed to \code{stats::optim(..., control = optimPass)}
if FALSE then list of control parameters will be inferred from other parameters.}

\item{stepsize}{Only for \code{IRLS}, scaling of updates to \code{beta} vector
lower value means slower convergence but more accuracy by default 1.
In general if fitting algorithm fails lowering this value tends to
be most effective at correcting it.}

\item{checkDiagWeights}{Logical value indicating whether to check if diagonal
elements of working weights matrixes in \code{IRLS} are sufficiently positive
so that these matrixes are positive defined. By default \code{TRUE}.}

\item{weightsEpsilon}{Small number to ensure positivity of weights matrixes.
Only matters if \code{checkDiagWeights} is set to \code{TRUE}.
By default \mjeqn{\approx 1.818989\cdot 10^{-12}}{approx. 1.818989 * 10^-12}}

\item{momentumFactor}{Experimental parameter in \code{IRLS} only allowing for
taking previous step into account at current step, i.e instead of
updating regression parameters as:
\mjdeqn{\boldsymbol{\beta}_{(a)} = \boldsymbol{\beta}_{(a-1)} + \text{stepsize} 
\cdot \text{step}_{(a)}}{beta_a = beta_a-1 + stepsize * step_a}
the update will be made as:
\mjdeqn{\boldsymbol{\beta}_{(a)} = \boldsymbol{\beta}_{(a-1)} + \text{stepsize} 
\cdot (\text{step}_{(a)} + \text{momentum}\cdot\text{step}_{(a-1)})}{
beta_a = beta_a-1 + stepsize * (step_a + momentum * step_a-1)}}

\item{saveIRLSlogs}{Logical value indicating if information specified in
\code{verbose} should be saved to output object, by default \code{FALSE}.}

\item{useZtpoissonAsStart}{boolean value indicating whether to chose starting
parameters from \code{ztpoisson} regression this one is useful mostly for
various one inflated models.}

\item{momentumActivation}{the value of log-likelihood reduction bellow
which momentum will apply.}

\item{criterion}{Criterion used to determine convergence in \code{IRLS},
multiple values may be provided. By default \code{c("coef", "abstol")}.}
}
\value{
List with selected parameters, it is also possible to call list directly.
}
\description{
\code{controlMethod} constructs a list with all necessary control parameters
for regression fitting in \code{estimatePopsize.fit} and \code{estimatePopsize}.
}
\seealso{
\code{\link[=estimatePopsize]{estimatePopsize()}} \code{\link[=estimatePopsize.fit]{estimatePopsize.fit()}} \code{\link[=controlModel]{controlModel()}} \code{\link[=controlPopVar]{controlPopVar()}}
}
\author{
Piotr Chlebicki, Maciej Beręsewicz
\loadmathjax
}
