% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{dfbeta.singleR}
\alias{dfbeta.singleR}
\title{TODO}
\usage{
\method{dfbeta}{singleR}(model, ...)
}
\arguments{
\item{model}{TODO}

\item{...}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\details{
dfbeta for singleRclass
}
