% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{dfbetasingleR}
\alias{dfbetasingleR}
\title{TODO}
\usage{
dfbetasingleR(model, method = c("formula", "simulation"), maxit.new = 1, ...)
}
\arguments{
\item{model}{Fitted object of singleR class}

\item{method}{Method for finding dfbetas, either "formula" where the formula for
approximate difference will be used or "simulation" where for each observation new
object will be fitted, with already obtained coefficients as starting values and with
just 1 iteration (or other number specified in maxit.new parameter), with i'th row
removed for all observations.}

\item{maxit.new}{maximal number of iterations for regression}

\item{...}{Arguments to pass to other methods}
}
\value{
TODO
}
\description{
TODO
}
\details{
dfbeta for singleRclass
}
