% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleRfit.R
\name{estimate_popsize.fit}
\alias{estimate_popsize.fit}
\title{Function to fit single-source capture-recapture models}
\usage{
estimate_popsize.fit(
  y,
  X,
  family,
  hwm,
  control,
  method,
  prior.weights,
  start,
  ...
)
}
\arguments{
\item{y}{Vector of dependent variables.}

\item{X}{A model matrix.}

\item{family}{Same as model in \code{estimate_popsize}.}

\item{control}{Control parameters.}
\item{hwm}{a vector containing information on how many columns in X matrix are associated with each linear predictor}

\item{control}{control parameters}

\item{method}{Method of estimation same as in \code{estimate_popsize}.}

\item{prior.weights}{Vector of weights its the same argument as weights
in \code{estimate_popsize}.}

\item{start}{Start for regression fitting.}

\item{dispersion}{Start for dispersion.}

\item{...}{Arguments to pass to other methods.}
\item{...}{arguments to pass to other methods}
}
\value{
list of object connected to regression
}
\description{
Function to fit single-source capture-recapture models
}
