% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{gridSearch.singleR}
\alias{gridSearch.singleR}
\title{TODO}
\usage{
gridSearch.singleR(object, updateCalculations = TRUE, ...)
}
\arguments{
\item{object}{TODO}

\item{updateCalculations}{TODO}

\item{...}{TODO}
}
\value{
TODO
}
\description{
TODO
}
