% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRLS.R
\name{IRLS}
\alias{IRLS}
\title{Iteratively Reweighted Least Squares}
\usage{
IRLS(
  dependent,
  family,
  covariates,
  start,
  disp = NULL,
  weights = NULL,
  maxiter = 10000,
  eps = .Machine$double.eps,
  disp.given = FALSE,
  silent = FALSE
)
}
\arguments{
\item{dependent}{A vector of Dependent variables}

\item{family}{Family of distributions used in regression}

\item{covariates}{An array of Covariates}

\item{start}{A place to start the numerical estimation}

\item{disp}{dispersion parameter if family needs it}

\item{weights}{Optional object of weights used in fitting the model}

\item{maxiter}{Maximum number of iterations}

\item{eps}{Precision level}

\item{disp.given}{FALSE is default, set true if dispersion has already been
estimated and there is no need for further estimation}

\item{silent}{Boolean value indicating whether warnings should be suppressed}
}
\value{
Returns a list of fitted Coefficients and number of iterations and final weights
}
\description{
A method for fitting regression parameters in generalized linear model. \cr
Used internally in singleRcapture package.
}
