% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.singleR}
\alias{plot.singleR}
\title{Diagnostic plots singleR regression}
\usage{
\method{plot}{singleR}(
  x,
  plotType = c("QQ", "marginal", "fitresid", "bootHist", "rootogram", "dfpopContr",
    "dfpopBox", "scaleLoc", "Cooks", "hatplot"),
  ...
)
}
\arguments{
\item{x}{TODO}

\item{plotType}{TODO}

\item{...}{TODO}
}
\value{
TODO
}
\description{
Diagnostic plots singleR regression
}
