% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoissonPopEst.R
\name{PoissonPopEst}
\alias{PoissonPopEst}
\title{Total Population Point and Interval estimate}
\usage{
PoissonPopEst(y, X, Grad, lambda, beta, Hess)
}
\arguments{
\item{y}{Observed values}

\item{X}{A covariate matrix}

\item{Grad}{A gradient of a model with respect to regression parameters}

\item{lambda}{A lambda Parameter for the model}

\item{beta}{Fitted regression values}

\item{Hess}{Hessian of a model}
}
\value{
Returns a list of size 3 with: \cr
(1) Point estimate \cr
(2) Variance \cr
(3) Confidence interval constructed using Point estimate plus minus 1.96 times standard deviation
}
\description{
Creates a point and interval estimate for total and missing population from zero truncated poisson model.
}
