% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratioReg.R, R/smallMethods.R
\name{popSizeEst.singleRRatioReg}
\alias{popSizeEst.singleRRatioReg}
\alias{popSizeEst}
\alias{popSizeEst.singleRStaticCountData}
\title{Extract population size estimation results}
\usage{
\method{popSizeEst}{singleRRatioReg}(object, estimator = c("primary", "HT", "SM"), ...)

popSizeEst(object, ...)

\method{popSizeEst}{singleRStaticCountData}(object, ...)
}
\arguments{
\item{object}{object with population size estimates.}

\item{estimator}{for \code{singleRRatioReg} objects, which population-size
estimator to return. The default \code{"primary"} returns the estimator
selected when calling \code{\link[=ratioReg]{ratioReg()}}.}

\item{...}{additional optional arguments, currently not used in \code{singleRStaticCountData} class method.}
}
\value{
An object of class \code{popSizeEstResults} containing population size estimation results.
}
\description{
An extractor function with \code{singleRStaticCountData} method for extracting
important information regarding pop size estimate.
}
