% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{print.summarysingleRmargin}
\alias{print.summarysingleRmargin}
\title{Print method for summarysingleRmargin classs}
\usage{
\method{print}{summarysingleRmargin}(x, ...)
}
\arguments{
\item{x}{object of class summarysingleRmargin}

\item{...}{Currently does nothing}
}
\value{
print of tests done by summary.singleRmargin
}
\description{
Print method for summarysingleRmargin classs
}
