% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{redoPopEstimation}
\alias{redoPopEstimation}
\alias{redoPopEstimation.singleR}
\title{Updating population size estimation results.}
\usage{
redoPopEstimation(object, ...)

\method{redoPopEstimation}{singleR}(object, cov = NULL, ...)
}
\arguments{
\item{object}{Object of singleR class.}

\item{...}{TODO}

\item{cov}{TODO}
}
\value{
TODO
}
\description{
A function
}
\examples{
# Create simple model
Model <- estimate_popsize(formula = capture ~ nation + age + gender, 
data = netherlandsimmigrant, 
model = ztpoisson, 
method = "robust")
# Apply heteroscedasticity consistent covariance matrix estimation
require(sandwich)
cov <- vcovHC(Model, type = "HC3")
summary(Model, cov = cov,
popSizeEst = redoPopEstimation(Model, cov = cov))
# Compare to results with usual covariance matrix estimation
summary(Model)
}
