% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{residuals.singleR}
\alias{residuals.singleR}
\title{Residuals of regresion}
\usage{
\method{residuals}{singleR}(object, type = c("pearson", "response", "working", "partial"), ...)
}
\arguments{
\item{object}{TODO}

\item{type}{TODO}

\item{...}{TODO}
}
\value{
returns a vector of residuals of
selected type
}
\description{
Residuals of regresion
}
\details{
S3 method for singleR class
}
