% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{simulate.singleR}
\alias{simulate.singleR}
\title{simulate}
\usage{
\method{simulate}{singleR}(object, nsim = 1, seed = NULL, ...)
}
\arguments{
\item{object}{an object representing a fitted model.}

\item{nsim}{number of response vectors to simulate. Defaults to \code{1}.}

\item{seed}{an object specifying if and how the random number generator should be initialized (‘seeded’).}

\item{...}{additional optional arguments.}
}
\value{
a \code{data.frame} with \code{n} rows and \code{nsim} columns.
}
\description{
An S3class for \code{stats::simulate} to handle \code{singleR} objects.
}
\examples{
set.seed(1)
N <- 10000
gender <- rbinom(N, 1, 0.2)
eta <- -1 + 0.5*gender
counts <- rpois(N, lambda = exp(eta))
df <- data.frame(gender, eta, counts)
df2 <- subset(df, counts > 0)
mod1 <-  estimate_popsize(formula = counts ~ 1 + gender, data = df2, 
model = "ztpoisson", method = "mle", pop.var = "analytic")
mod1_sims <- simulate(mod1, nsim=10)
colMeans(mod1_sims) 
}
\seealso{
\code{\link[stats:simulate]{stats::simulate()}}
}
