% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleRmodel.R
\name{Single_Rmodel}
\alias{Single_Rmodel}
\title{SingleRmodel}
\usage{
Single_Rmodel(
  data,
  formula,
  start = NULL,
  dispersion = NULL,
  family = stop("Family argument is needed"),
  Fit.Method = "IRLS",
  CI = "analytic",
  prior.weights = 1,
  ...
)
}
\arguments{
\item{data}{Data frame for the regression}

\item{formula}{Description of model to that is supposed to be fitted}

\item{start}{Optional argument specifying initial values of regression parameters}

\item{dispersion}{optional parameter if family exhibits over/underdispersion
and it has already known/been estimated providing it should improve model fit considerably}

\item{family}{Family of distributions used in regression}

\item{Fit.Method}{Method for fitting values currently supported IRLS and MaxLikelihood}

\item{CI}{A method of constructing confidence interval either analytic or bootstrap
where bootstraped confidence interval may either be based on 2.5\%-97.5\%
percientiles ("bootstrapPerc") or studentized CI ("bootstrapSD")}

\item{prior.weights}{Optional object of a-priori weights used in fitting the model}

\item{...}{Arguments to be passed to other methods like subset in stats::model.frame}
}
\value{
Returns an object of classes inherited from glm containing:\cr

y argument used in regression\cr
formula provided on call\cr
Call on which model is based\cr
vector of fitted coefficients of regression\cr
vector of Standard errors of coeffcients\cr
vectors of test and p - values \cr
family used in regression\cr
akane information criterion\cr
vector of prior weights assingned on call\cr
vector of weights if IRLS was the fit method then its a vector what was returned by IRLS\cr
vector of working residuals\cr
Log-likelihood for the model\cr
number of itteration performed in fitting\cr
total and reduced degrees of freedom\cr
vector of fitted values\cr
a list containing population size estimate form PopulationEstimate function\cr
model, linear predictors, and qr decomposition of model matrix
}
\description{
SingleRmodel
}
