% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleRmodel.R
\name{SingleRmodel}
\alias{SingleRmodel}
\title{SingleRmodel}
\usage{
SingleRmodel(
  data,
  Dependent,
  Explanatory,
  Family = stop("Family is needed"),
  Fit.Method = "IRLS"
)
}
\arguments{
\item{data}{Data frame for the regression}

\item{Dependent}{Name of column of dependent variables in data}

\item{Explanatory}{Vector of names of columns of explanatory variables (covariates) in data}

\item{Family}{Family of distributions used in regression}

\item{Fit.Method}{Method for fitting values currently supported IRLS and MaxLikelihood}
}
\value{
Returns a S3 object with:\cr fitted values\cr standard errors\cr akane information criterion \cr
Tvalues and Pvalues for testing the hypothesis of no effect of coefficient on expected value \cr
Names of used variables\cr response residuals\cr estimate of population size pearson residuals and parameter\cr
Estimation of true population size
}
\description{
SingleRmodel
}
