% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryModel.R
\name{summary.Model}
\alias{summary.Model}
\title{summary.Model}
\usage{
\method{summary}{Model}(object)
}
\arguments{
\item{object}{Object to be summarised}
}
\value{
Easy to read summary of regression and most important data from a 'Model' class
}
\description{
Describes an object of model class representing generalized linear regression
}
