% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.SingleR}
\alias{summary.SingleR}
\title{summary.Model}
\usage{
\method{summary}{SingleR}(object, ...)
}
\arguments{
\item{object}{Object to be summarised}

\item{...}{other arguments to be passed to other methods}
}
\value{
Easy to read summary of regression and most important data from a 'SingleR' class
}
\description{
Describes an object of model class representing generalized linear regression
}
