% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratioReg.R
\name{summary.singleRRatioReg}
\alias{summary.singleRRatioReg}
\title{Summary Statistics for Ratio Regression Fits}
\usage{
\method{summary}{singleRRatioReg}(object, ...)
}
\arguments{
\item{object}{object of class \code{singleRRatioReg}.}

\item{...}{currently not used.}
}
\value{
An object of class \code{summarysingleRRatioReg}.
}
\description{
Summarizes a fitted \code{singleRRatioReg} object, including model-selection
criteria and population-size estimates.
}
