% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightedSd.R
\name{Weighted_Sd}
\alias{Weighted_Sd}
\title{Weighted_sd}
\usage{
Weighted_Sd(x, w, ...)
}
\arguments{
\item{x}{A vector of values from which standard deviation is to be calculated.}

\item{w}{A vector of weights.}

\item{...}{Arguments to be passed from/to other methods.}
}
\value{
numeric vector of weighted standard deviation
}
\description{
Weighted_sd
}
