% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZeroTruncatedNegativeBinomial.R
\name{Zero_Truncated_Negative_Binomial}
\alias{Zero_Truncated_Negative_Binomial}
\title{ZeroTruncatedNegativeBinomial}
\usage{
Zero_Truncated_Negative_Binomial()
}
\value{
A list of
}
\description{
ZeroTruncatedNegativeBinomial
}
