% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZeroTruncatedPoisson.R
\name{Zero_Truncated_Poisson}
\alias{Zero_Truncated_Poisson}
\title{Zero Truncated Poisson Model Functions}
\usage{
Zero_Truncated_Poisson()
}
\value{
A list of objects \cr
make_minusloglike(y,X) - for creating negative likelihood function \cr
make_gradient(y,X) - for creating gradient function \cr
make_hessian(X) - for creating hessian \cr
link - a link function to connect between linear predictor and model parameter in regression \cr
Invlink - an inverse function of link \cr
Dlink - a 1sr derivative of link function \cr
mu,Variance - Expected Value and Variance for full model (without truncation) \cr
Variance - Variance for full model (without truncation) multiplied by Dispersion parameter \cr
Where: \cr
y is a vector of observed values \cr
X is a matrix / data frame of covariates
}
\description{
Used internally withing a package creates gradient hessian and minus log likelihood
for Zero Truncated Poisson
}
