% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZeroTruncatedPoisson.R
\name{ZeroTruncatedPoisson}
\alias{ZeroTruncatedPoisson}
\title{Zero Truncated Poisson Model Functions}
\usage{
ZeroTruncatedPoisson()
}
\value{
A list of 3 objects \cr
make_minusloglike(y,X) - for creating negative likelihood function \cr
make_gradient(y,X) - for creating gradient function \cr
make_hessian(X) - for creating hessian \cr
Where: \cr
y is a vector of observed values \cr
X is a matrix / data frame of covariates
}
\description{
Used internally withing a package creates gradient hessian and minus log likelihood
for Zero Truncated Poisson
}
