% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zotgeom.R
\name{zotgeom}
\alias{zotgeom}
\title{Zero-one truncated geometric model}
\usage{
zotgeom()
}
\value{
A object of class "family" containing objects \cr
makeMinusLogLike(y,X) - for creating negative likelihood function \cr
makeGradient(y,X) - for creating gradient function \cr
makeHessian(X) - for creating hessian \cr
linkfun - a link function to connect between linear predictor and model parameter in regression and a name of link function\cr
linkinv - an inverse function of link \cr
Dlink - a 1st derivative of link function \cr
mu.eta,Variance - Expected Value and Variance \cr
valedmu, valideta - for checking if regression arguments and valid\cr
family - family name\cr
Where: \cr
y is a vector of observed values \cr
X is a matrix / data frame of covariates
}
\description{
Zero-one truncated geometric model
}
