% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZeroOneTruncatedNegativeBinomial.R
\name{zotnegbin}
\alias{zotnegbin}
\title{Zero One Truncated Negative Binomial model}
\usage{
zotnegbin(nSim = 1000, epsSim = 1e-08, ...)
}
\arguments{
\item{nSim}{TODO}

\item{epsSim}{TODO}
}
\value{
A object of class "family" containing objects \cr
makeMinusLogLike(y,X) - for creating negative likelihood function \cr
makeGradient(y,X) - for creating gradient function \cr
makeHessian(X) - for creating hessian \cr
linkfun - a link function to connect between linear predictor and model parameter in regression and a name of link function\cr
linkinv - an inverse function of link \cr
Dlink - a 1st derivative of link function \cr
mu.eta,Variance - Expected Value and Variance \cr
valedmu, valideta - for checking if regression arguments and valid\cr
family - family name\cr
Where: \cr
y is a vector of observed values \cr
X is a matrix / data frame of covariates
}
\description{
Zero One Truncated Negative Binomial model
}
