% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chao.R, R/hurdlegeom.R, R/hurdlegeom2.R,
%   R/hurdlepoisson.R, R/hurdlepoisson2.R, R/oiztgeom.R, R/oiztpoisson.R,
%   R/zelterman.R, R/zotgeom.R, R/zotnegbin.R, R/zotpoisson.R, R/ztgeom.R,
%   R/ztnegbin.R, R/ztoigeom.R, R/ztoipoisson.R, R/ztpoisson.R
\name{chao}
\alias{chao}
\alias{ztHurdlegeom}
\alias{Hurdleztgeom}
\alias{ztHurdlepoisson}
\alias{Hurdleztpoisson}
\alias{oiztgeom}
\alias{oiztpoisson}
\alias{zelterman}
\alias{zotgeom}
\alias{zotnegbin}
\alias{zotpoisson}
\alias{ztgeom}
\alias{ztnegbin}
\alias{ztoigeom}
\alias{ztoipoisson}
\alias{ztpoisson}
\title{Family functions in singleRcapture package}
\usage{
chao(...)

ztHurdlegeom(...)

Hurdleztgeom(...)

ztHurdlepoisson(...)

Hurdleztpoisson(...)

oiztgeom(...)

oiztpoisson(...)

zelterman(...)

zotgeom(...)

zotnegbin(nSim = 1000, epsSim = 1e-08, ...)

zotpoisson(...)

ztgeom(...)

ztnegbin(nSim = 1000, epsSim = 1e-08, ...)

ztoigeom(...)

ztoipoisson(...)

ztpoisson(...)
}
\arguments{
\item{...}{Additional arguments, not used for now.}

\item{nSim}{If working weights cannot be computed analytically this argument
specifies maximum number of simulations allowed to find them numerically.}

\item{epsSim}{f working weights cannot be computed analytically this argument
specifies precision level for finding them numerically.}
}
\value{
A object of class \code{family} containing objects:
\itemize{
\item \code{makeMinusLogLike, makeGradient, makeHessian} -- A factory functions for creating
\mjeqn{\ell(\boldsymbol{\beta}), \frac{\partial\ell}{\partial\boldsymbol{\beta}},
\frac{\partial^{2}\ell}{\partial\boldsymbol{\beta}}^{T}\partial\boldsymbol{\beta}
}{l(beta), dl/dbeta, d^2l/dbeta^Tdbeta} functions respectively from
\mjeqn{\boldsymbol{y}}{y} vector and \mjeqn{\boldsymbol{X}_{vlm}}{X_vlm}
(or just \mjeqn{\boldsymbol{X}}{X} if applied to model with single linear predictor).
\item \code{linkfun, linkinv} -- A link function and its inverse.
\item \code{mu.eta, variance} -- Functions of linear predictors that
return expected value and variance. There is a \code{type} argument with
2 possible values \code{"trunc"} and \code{"nontrunc"} that specifies whether
to return \mjeqn{\mathbb{E}(Y|Y>0), \text{var}(Y|Y>0}{E(Y|Y>0), var(Y|Y>0)}
or \mjeqn{\mathbb{E}Y, \text{var}(Y)}{E(Y), var(Y)} respectively.
\item \code{link, family} -- Character specifying link functions used for the
model and the internally used name of the model.
\item \code{valideta, validmu} -- For now only returns true. In near future
will be used to check whether applied linear predictors are valid (i.e. are
transormed into some elements of parameter space the subjected to inverse
link function).
\item \code{funcZ, Wfun} -- Functions that create pseudo residuals and
working weights used in IRLS algorithm.
\item \code{dev.resids} -- Function that given the linear predictors
prior weights vector and response vector returns deviance residuals.
Not all family functions have these functions implemented yet.
\item \code{pointEst, popVar} -- Functions that given prior weights
linear predictors and in the later case also estimation of
\mjeqn{\text{cov}(\hat{\boldsymbol{\beta}})}{cov(beta)}
and \mjeqn{\boldsymbol{X_vlm}}{X_vlm} matrix return point estimate
for population size and analytic estimation of its variance.
There is a additional boolean parameter \code{contr} in the former
function that if set to true returns contribution of each unit.
\item \code{etaNames, parNum} -- Names and number of linear predictors.
\item \code{densityFunction} -- A function that given linear predictors
returns value of PMF at values \code{x}. Additional argument \code{type}
spedifies whether to return \mjeqn{\mathbb{P}(Y|Y>0)}{P(Y|Y>0)} or
\mjeqn{\mathbb{P}(Y)}{P(Y)}.
\item \code{simulate} -- A function that generates values of dependent
vector given linear predictors.
}
}
\description{
Package \code{singleRcapture} utilises various family type
functions that specify variable parts of population size estimation,
regression, diagnostics and other necessary information that depends
on the model. These functions are used as \code{model} argument in
\code{estimate_popsize} function.
}
\details{
\loadmathjax

TODO important describe everything.
}
\seealso{
\code{\link[=estimate_popsize]{estimate_popsize()}}
}
\author{
Piotr Chlebicki, Maciej Beręsewicz
}
