% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{bread.singleR}
\alias{bread.singleR}
\title{bread}
\usage{
\method{bread}{singleR}(object, ...)
}
\arguments{
\item{object}{an object representing a fitted model.}

\item{...}{additional optional arguments.}
}
\value{
a
}
\description{
An S3class for \code{sandwich::bread} to handle \code{singleR} objects. This function was developed based on \code{sandwich:::bread.glm}
}
\seealso{
\code{\link[sandwich:bread]{sandwich::bread()}}
}
