% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{carcassubmission}
\alias{carcassubmission}
\title{Carcass submissions}
\format{
Data frame with 1858 rows and 4 columns.
\describe{
\item{TOTAL_SUB}{Number of submissions}
\item{log_size}{logarithm of size of farm}
\item{log_distance}{logarith of distance to nearest AHVLA center}
\item{C_TYPE}{Type of farm 1 for dairy and 0 for beef}
}
}
\usage{
data(carcassubmission)
}
\description{
Carcass submissions
}
\details{
Data on british animal farms submissions to AHVLA. It is required
for british farms to submit samples to AHVLA if unknown disease
if private veterinary surgeons cannot determine the cause of
death for the animal and decide to submit them, unless there is notifiable
disease suspected then it is required. This is data on such farms. This
data set contains farms that have submitted animal carcasses.
}
\references{
Bohning et.al (2013) BIOMETRICS 69, 1033-1042
}
\keyword{datasets}
