% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{confint.singleR}
\alias{confint.singleR}
\title{Confidence Intervals for Model Parameters}
\usage{
\method{confint}{singleR}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{object of singleR class.}

\item{parm}{names of parameters for which confidence intervals are to be
computed, if missing all parameters will be considered.}

\item{level}{confidence level for intervals.}

\item{...}{currently does nothing.}
}
\value{
An object with named columns that include upper and
lower limit of confidence intervals.
}
\description{
A function that computes studentized confidence intervals
for model coefficients.
}
