% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{control.method}
\alias{control.method}
\title{Control parameters for regression}
\usage{
control.method(
  epsilon = 1e-08,
  maxiter = 1000,
  verbose = 0,
  start = NULL,
  alphaStart = NULL,
  omegaStart = NULL,
  piStart = NULL,
  mleMethod = "L-BFGS-B",
  silent = FALSE,
  optimPass = FALSE,
  stepsize = 1,
  momentumFactor = 0,
  momentumActivation = 5
)
}
\arguments{
\item{epsilon}{relative tolerance (absolute for robust method) for fitting algorithms by default 1e-8.}

\item{maxiter}{Maximal number of iterations.}

\item{verbose}{Value indicating whether to trace steps of fitting algorithm for robust fitting method different values of verbose give the following information:
\itemize{
\item 1 -- Returns information on the number of current iteration and current log-likelihood.
\item 2 -- Returns information on vector of regression parameters at current iteration (and all of the above).
\item 3 -- Returns information on reduction of log-likelihood at current iteration (and all of the above).
\item 4 -- Returns information on value of log-likelihood function gradient at current iteration (and all of the above).
\item 5 -- Returns information on convergence criterion and values that are taken into account when considering convergence (and all of the above).
}
if mle method was chosen verbose will be passed to \code{\link[stats:optim]{stats::optim()}} as trace.}

\item{start}{initial parameters for regression associated with main formula specified in function call if NULL they will be derived from simple poisson regression.}

\item{alphaStart}{initial parameters for dispersion parameter if applies.}

\item{omegaStart}{initial parameters for inflation parameter if applies.}

\item{piStart}{initial parameters for probability parameter if applies.}

\item{mleMethod}{method of \code{\link[stats:optim]{stats::optim()}} used L-BFGS-B is the default except for negative binomial and one inflated models where Nelder-Mead is used.}

\item{silent}{logical indicating whether warnings in robust method should be suppressed.}

\item{optimPass}{optional parameter allowing for passing list into as stats::optim(..., control = optimPass) if FALSE then list of control parameters will be inferred from other parameters.}

\item{stepsize}{only for robust, scaling of stepsize lower value means slower convergence but more accuracy by default 1. In general if fitting algorithm fails lowering this value tends to be most effective at correcting it.}

\item{momentumFactor}{experimental parameter in robust only allowing for taking previous step into account at current step, i.e instead of updating regression parameters as:
\loadmathjax
\mjsdeqn{\boldsymbol{\beta}_{(a)} = \boldsymbol{\beta}_{(a-1)} + \text{stepsize} \cdot \text{step}_{(a)}}
the update will be made as:
\mjsdeqn{\boldsymbol{\beta}_{(a)} = \boldsymbol{\beta}_{(a-1)} + \text{stepsize} \cdot (\text{step}_{(a)} + \text{step}_{(a-1)})}}

\item{momentumActivation}{the value of log-likelihood reduction bellow which momentum will apply.}
}
\value{
A list with selected parameters, it is also possible to call list directly.
}
\description{
Control parameters for regression
}
