% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{control.method}
\alias{control.method}
\title{Control parameters for regression}
\usage{
control.method(
  epsilon = .Machine$double.eps,
  maxiter = 1000,
  trace = "no",
  start = NULL,
  dispersionstart = NULL,
  mleMethod = "L-BFGS-B",
  silent = FALSE,
  optimPass = FALSE,
  disp.given = FALSE
)
}
\arguments{
\item{epsilon}{relative tolerance for fitting algorithms}

\item{maxiter}{Maximal number of iterations}

\item{trace}{Value indicating whether to trace steps of fitting algorithm for robust its either "no", "logL" or "beta", for mle it should be logical}

\item{start}{initial parameters for regression if NULL they will be derived from simple poisson regression}

\item{dispersionstart}{initial parameters for dispersion parameter if applies}

\item{mleMethod}{method of stats::optim used L-BFGS-B is the default except for negative binomial models where Nelder-Mead is used}

\item{silent}{logical indicating whether warnings in robust method should be suppressed}

\item{optimPass}{optional parameter allowing for passing list into as stats::optim(..., control = optimPass) if FALSE then list of controlparameters will be infered from other parameters}

\item{disp.given}{logical indicates whether to estimate dispersion parameter or assume the one provided is already correct}
}
\value{
A list with selected parameters, it is also possible to call list directly.
}
\description{
Control parameters for regression
}
