% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{control.method}
\alias{control.method}
\title{Control parameters for regression}
\usage{
control.method(
  epsilon = 1e-08,
  dispEpsilon = 1e-05,
  maxiter = 1000,
  verbose = 0,
  start = NULL,
  dispersionstart = NULL,
  omegastart = NULL,
  mleMethod = "L-BFGS-B",
  silent = FALSE,
  optimPass = FALSE,
  dispGiven = FALSE,
  thetaGiven = FALSE,
  stepsize = 1
)
}
\arguments{
\item{epsilon}{relative tolerance (absolute for robust method) for fitting algorithms by default 1e-8}

\item{dispEpsilon}{tolerance for estimation of dispersion parameter only matters for negbins in robust}

\item{maxiter}{Maximal number of iterations}

\item{verbose}{Value indicating whether to trace steps of fitting algorithm for robust its either 0 (for no tracing), 1 (for tracing logarithm likelihhod) or 2 (for tracing logarithm likelihood and vector of regression parameters) for mle it is passed to  stats::optim as trace}

\item{start}{initial parameters for regression if NULL they will be derived from simple poisson regression}

\item{dispersionstart}{initial parameters for dispersion parameter if applies}

\item{omegastart}{initial parameters for inflation parameter if applies}

\item{mleMethod}{method of stats::optim used L-BFGS-B is the default except for negative binomial models where Nelder-Mead is used}

\item{silent}{logical indicating whether warnings in robust method should be suppressed}

\item{optimPass}{optional parameter allowing for passing list into as stats::optim(..., control = optimPass) if FALSE then list of controlparameters will be infered from other parameters}

\item{dispGiven}{logical indicates whether to estimate dispersion parameter or assume the one provided is already correct}

\item{thetaGiven}{TODO}

\item{stepsize}{Only for robust, scaling of stepsize lower value means slower convergence but more accuracy by default 1}
}
\value{
A list with selected parameters, it is also possible to call list directly.
}
\description{
Control parameters for regression
}
