% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{control.method}
\alias{control.method}
\title{Control parameters for regression}
\usage{
control.method(
  epsilon = 1e-08,
  dispEpsilon = 1e-05,
  maxiter = 1000,
  verbose = 0,
  start = NULL,
  dispersionstart = NULL,
  mleMethod = "L-BFGS-B",
  silent = FALSE,
  optimPass = FALSE,
  disp.given = FALSE,
  stepsize = 1
)
}
\arguments{
\item{epsilon}{Relative tolerance for fitting algorithms by default 1e-8.}

\item{maxiter}{Maximum number of iterations.}

\item{verbose}{Value indicating whether to trace steps of fitting algorithm for robust its either 0 (for no tracing), 1 (for tracing logarithm likelihood) or 2 (for tracing logarithm likelihood and vector of regression parameters) for mle it is passed to  stats::optim as trace.}

\item{start}{Initial parameters for regression coefficients if NULL they will be derived from simple poisson regression.}

\item{dispersionstart}{Initial parameters for dispersion parameter if applicable.}

\item{mleMethod}{Passed to stats::optim. L-BFGS-B is the default except for negative binomial models where Nelder-Mead is used.}

\item{silent}{Logical, indicating whether warnings in robust method should be suppressed.}

\item{optimPass}{Optional list of parameters passed to stats::optim(..., control = optimPass) if FALSE then list of control parameters will be inferred from other parameters.}

\item{disp.given}{Logical, indicates whether to estimate dispersion parameter or assume the one provided is already correct.}

\item{stepsize}{Only for robust, scaling of stepsize lower value means slower convergence but more accuracy by default 1.}
}
\value{
List with selected parameters, it is also possible to call list directly.
}
\description{
Control parameters for regression
}
