% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{control.model}
\alias{control.model}
\title{control.model
TODO}
\usage{
control.model(
  weightsAsCounts = FALSE,
  omegaFormula = NULL,
  alphaFormula = NULL
)
}
\arguments{
\item{weightsAsCounts}{TODO}

\item{omegaFormula}{TODO}

\item{alphaFormula}{TODO
for now does nothing}
}
\value{
control.model
}
\description{
control.model
TODO
}
