% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{control.pop.var}
\alias{control.pop.var}
\title{Control parameters for population size estimation}
\usage{
control.pop.var(
  alpha = 0.05,
  trcount = 0,
  bootType = c("parametric", "semiparametric", "nonparametric"),
  B = 500,
  confType = c("percentilic", "studentized", "basic"),
  keepbootStat = TRUE,
  traceBootstrapSize = FALSE,
  fittingMethod = NULL,
  bootstrapFitcontrol = NULL,
  covType = c("observedInform", "Fisher")
)
}
\arguments{
\item{alpha}{Significance level, 0.05 used by default.}

\item{trcount}{Truncated count -- a number to be added to point estimator and both sides of confidence intervals.}

\item{bootType}{Bootstrap type. Default is \code{"parametric"}, other possible values are: \code{"semiparametric"} and \code{"nonparametric"}.}

\item{B}{Number of bootstrap samples to be performed (default 500).}

\item{confType}{Type of confidence interval for bootstrap confidence interval, \code{"percentile"} by default. Other possibility: \code{"studentized"} and \code{"basic"}.}

\item{keepbootStat}{Boolean value indicating whether to keep a vector of statistics produced by bootstrap.}

\item{traceBootstrapSize}{Boolean value indicating whether to print size of bootstrapped sample after truncation for semi- and fully parametric bootstraps.}

\item{fittingMethod}{Method used for fitting models from bootstrap samples.}

\item{bootstrapFitcontrol}{Control parameters for each regression works exactly like \code{control.method}}

\item{covType}{Type of covariance matrix for regression parameters by default observed information matrix, more options will be here in the future.}
}
\value{
A list with selected parameters, it is also possible to call list directly.
}
\description{
Control parameters for population size estimation
}
