% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{control.pop.var}
\alias{control.pop.var}
\title{Control parameters for population size estimation}
\usage{
control.pop.var(
  alpha = 0.05,
  trcount = 0,
  bootType = c("parametric", "semiparametric", "nonparametric"),
  B = 500,
  confType = c("percentilic", "studentized", "basic"),
  keepbootStat = TRUE,
  traceBootstrapSize = FALSE,
  fittingMethod = NULL,
  bootstrapFitcontrol = NULL,
  sd = c("sqrtVar", "normalMVUE"),
  covType = c("observedInform", "Fisher")
)
}
\arguments{
\item{alpha}{Significance level, a number from range (0, 1) 5\% by default.}

\item{trcount}{Truncated count - a number to be added to point estimator and both sides of confidence intervals.}

\item{bootType}{Type of bootstrap performed, by default Parametric, other posible values are: Semiparametric and Nonparametric.}

\item{B}{Number of bootstrap samples to be performed.}

\item{confType}{Type of confidence interval for bootstrap confidence interval, percentilic by default.}

\item{keepbootStat}{Boolean value indicating whether to keep a vector of statistic produced by bootstrap, for large values of B it may significant amount of size.}

\item{traceBootstrapSize}{Boolean value indicating whether to print size of bootstrapped sample after truncation for semi and fully parametric boostraps.}

\item{fittingMethod}{Method used for fitting models from boostrap samples either "robust" or "mle" is left as NULL will be chosen automatically.}

\item{bootstrapFitcontrol}{Control parameters for each regression works exactly like control.method.}

\item{sd}{Indicates how to compute standard deviation of population size estimator either as:
\mjdeqn{\hat{\sigma}=\sqrt{\hat{\text{var}}(\hat{N})}}{sd=sqrt(var(N))}
for sqrt or for normalMVUE as the unbiased minimal variance estimator for normal distribution:
\mjdeqn{\hat{\sigma}=\sqrt{\hat{\text{var}}(\hat{N})}\frac{\Gamma\left((N_{obs}-1)/2\right)}{\Gamma\left(N_{obs}/2\right)}\sqrt{\frac{N_{obs}}{2}}}{sd=sqrt(var(N))sqrt(N_obs/2)Gamma(N_obs-1/2)/Gamma(N_obs/2)}
where the ration involving gamma functions is computed by loggamma function.}

\item{covType}{type of covariance matrix for regression parameters by default observed information matrix, more options will be here in the future.}
}
\value{
A list with selected parameters, it is also possible to call list directly.
}
\description{
\loadmathjax
}
