% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{control.pop.var}
\alias{control.pop.var}
\title{Control parameters for variance estimation}
\usage{
control.pop.var(
  signiflevel = 0.95,
  trcount = 0,
  bootType = "Parametric",
  strapNumber = 500,
  confType = "Percentilic"
)
}
\arguments{
\item{signiflevel}{Significance level, a number from frange (0, 1) 95\% by default.}

\item{trcount}{Truncated count - a number to be added to point estimator and both sides of confidence intervals.}

\item{bootType}{Type of bootstrap performed, by default Parametric, other posible values are: Semiparametric and Nonparametric.}

\item{strapNumber}{Number of bootstrap resamplings to be performed.}

\item{confType}{a type of confidence interval for bootstrap confidence intervals, Percentilic by default, may be change to studentized.}
}
\value{
A list with selected parameters, it is also possible to call list directly
}
\description{
Control parameters for variance estimation
}
