% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{control.pop.var}
\alias{control.pop.var}
\title{Control parameters for population size estimation}
\usage{
control.pop.var(
  alpha = 0.05,
  trcount = 0,
  bootType = c("parametric", "semiparametric", "nonparametric"),
  B = 500,
  confType = c("percentilic", "studentized", "basic"),
  keepbootStat = TRUE,
  traceBootstrapSize = FALSE,
  bootstrapVisualTrace = FALSE,
  fittingMethod = NULL,
  bootstrapFitcontrol = NULL,
  sd = c("sqrtVar", "normalMVUE"),
  covType = c("observedInform", "Fisher")
)
}
\arguments{
\item{alpha}{Significance level, 0.05 used by default.}

\item{trcount}{Truncated count - a number to be added to point estimator and both sides of confidence intervals.}

\item{bootType}{Bootstrap type. Default is \code{"parametric"}, other possible values are: \code{"semiparametric"} and \code{"nonparametric"}.}

\item{B}{Number of bootstrap samples to be performed (default 500).}

\item{confType}{Type of confidence interval for bootstrap confidence interval, \code{"percentile"} by default. Other possibility: \code{"studentized"} and \code{"basic"}.}

\item{keepbootStat}{Boolean value indicating whether to keep a vector of statistics produced by bootstrap.}

\item{traceBootstrapSize}{Boolean value indicating whether to print size of bootstrapped sample after truncation for semi- and fully parametric bootstraps.}

\item{bootstrapVisualTrace}{Boolean value indicating whether to plot bootstrap statistics in real time.}

\item{fittingMethod}{Method used for fitting models from bootstrap samples.}

\item{bootstrapFitcontrol}{Control parameters for each regression works exactly like \code{control.method} but for fitting models from bootstrap samples.}

\item{sd}{Indicates how to compute standard deviation of population size estimator either as:
\loadmathjax
\mjdeqn{\hat{\sigma}=\sqrt{\hat{\text{var}}(\hat{N})}}{sd=sqrt(var(N))}
for sqrt or for normalMVUE as the unbiased minimal variance estimator for normal distribution:
\mjdeqn{\hat{\sigma}=\sqrt{\hat{\text{var}}(\hat{N})}\frac{\Gamma\left((N_{obs}-1)/2\right)}{\Gamma\left(N_{obs}/2\right)}\sqrt{\frac{N_{obs}}{2}}}{sd=sqrt(var(N))sqrt(N_obs/2)Gamma(N_obs-1/2)/Gamma(N_obs/2)}
where the ration involving gamma functions is computed by loggamma function.}

\item{covType}{type of covariance matrix for regression parameters by default observed information matrix, more options will be here in the future.}
}
\value{
A list with selected parameters, it is also possible to call list directly.
}
\description{
Creating control parameters for population size estimation and
respective standard error and variance estimation.
}
\seealso{
\code{\link[=estimate_popsize]{estimate_popsize()}} \code{\link[=control.model]{control.model()}} \code{\link[=control.method]{control.method()}}
}
\author{
Piotr Chlebicki, Maciej Beręsewicz
}
