% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{dfbetasingleR}
\alias{dfbetasingleR}
\title{TODO}
\usage{
dfbetasingleR(model, maxit.new = 1, ...)
}
\arguments{
\item{model}{Fitted object of singleR class}

\item{maxit.new}{maximal number of iterations for regression}

\item{...}{Arguments to pass to other methods}
}
\value{
TODO
}
\description{
TODO
}
\details{
dfbeta for singleRclass
}
