% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwichMethods.R
\name{estfun.singleR}
\alias{estfun.singleR}
\title{estfun}
\usage{
\method{estfun}{singleR}(object, ...)
}
\arguments{
\item{object}{an object representing a fitted model.}

\item{...}{additional optional arguments.}
}
\value{
a \code{matrix} with \code{n} rows and \code{k} columns where \code{k} denotes number of variables.
}
\description{
An S3class for \code{sandwich::estfun} to handle \code{singleR} objects. This function was developed based on \code{countreg::estfun.zerotrunc}
}
\examples{
set.seed(1)
N <- 10000
gender <- rbinom(N, 1, 0.2)
eta <- -1 + 0.5*gender
counts <- rpois(N, lambda = exp(eta))
df <- data.frame(gender, eta, counts)
df2 <- subset(df, counts > 0)
mod1 <-  estimate_popsize(formula = counts ~ 1 + gender, data = df2, 
model = "ztpoisson", method = "mle", pop.var = "analytic")
mod1_sims <- sandwich::estfun(mod1)
head(mod1_sims) 
}
\seealso{
\code{\link[sandwich:estfun]{sandwich::estfun()}}
}
