% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_popsize.fit.R
\name{estimate_popsize.fit}
\alias{estimate_popsize.fit}
\title{Regression fitting in single source capture-recapture models}
\usage{
estimate_popsize.fit(y, X, family, control, method, prior.weights, start, ...)
}
\arguments{
\item{y}{vector of dependent variables.}

\item{X}{model matrix, the vglm one.}

\item{family}{same as model in \code{estimate_popsize}.}

\item{control}{control parameters created in \code{control.model}.}

\item{method}{method of estimation same as in \code{estimate_popsize}.}

\item{prior.weights}{vector of prior weights its the same argument as weights
in \code{estimate_popsize}.}

\item{start}{initial value of regression parameters.}

\item{...}{arguments to pass to other methods.}
}
\value{
List with regression parameters, working weights
(if IRLS fitting method) was chosen and number of iterations taken.
}
\description{
\code{estimate_popsize.fit} does for \code{estimate_popsize} what
\code{glm.fit} does for \code{glm}. It is internally called in
\code{estimate_popsize}. Since \code{estimate_popsize} does much more than
just regression fitting \code{estimate_popsize.fit} is much faster.
}
\details{
If \code{method} argument was set to \code{"optim"} the \code{stats::optim}
function will be used to fit regression with analyticly computed gradient and
(minus) log likelihood functions as \code{gr} and \code{fn} arguments.
Unfortunately \code{optim} does not allow for hessian to be specified.
More information about how to modify \code{optim} fitting is included in
\code{\link[=control.method]{control.method()}}.

If \code{method} argument was set to \code{"IRLS"} the iteratively reweighted
least squares. The algorithm is well know in generalised linear models.
Thomas W. Yee later extended this algorithm to vector generalised linear models
and in more general terms it can roughly be described as
(this is Yee's description after changing some conventions):
\loadmathjax
\enumerate{
\item Initialise with:
\itemize{
\item \code{converged <- FALSE}
\item \code{iter <- 1}
\item \mjeqn{\boldsymbol{\beta}}{beta}\code{ <- start}
\item \mjeqn{\boldsymbol{W}}{W}\code{ <- prior}
\item \mjeqn{\ell}{l}\code{ <- }\mjeqn{\ell(\boldsymbol{\beta})}{l(beta)}
}
\item If \code{converged} or \code{iter > Maxiter} move to step 7.
\item Store values from previous algorithm step:
\itemize{
\item \mjeqn{\boldsymbol{W}_{-}}{W_-}\code{ <- }\mjeqn{\boldsymbol{W}}{W}
\item \mjeqn{\ell_{-}}{l_-}\code{ <- }\mjeqn{\ell}{l}
\item \mjeqn{\boldsymbol{\beta}_{-}}{beta_-}\code{ <- }\mjeqn{\boldsymbol{\beta}}{beta}
} and assign values at current step:
\itemize{
\item \mjeqn{\boldsymbol{\eta}}{eta}\code{ <- }\mjeqn{\boldsymbol{X}_{vlm}\boldsymbol{\beta}}{X_vlm * beta}
\item \mjeqn{Z_{i}}{Z_i}\code{ <- }\mjeqn{\eta_{i}+\frac{\partial\ell_{i}}{\partial\eta_{i}}\mathbb{E}\left(\frac{\partial^{2}\ell_{i}}{\partial\eta_{i}^{T}\partial\eta_{i}}\right)^{-1}}{eta_i + (dl/deta_i)(E(d^2l/deta_i^Tdeta_i))^-1}
\item \mjeqn{\boldsymbol{W}_{ij}}{W_ij}\code{ <- }\mjeqn{\mathbb{E}\left(\frac{\partial^{2}\ell}{\partial\boldsymbol{\eta}_{j}^{T}\partial\boldsymbol{\eta}_{i}}\right)}{E(d^2l/deta_j^Tdeta_i)}
}
where \mjeqn{\ell_{i}}{l_i} is the ith component of log likelihood function,
\mjeqn{\eta_{i}}{eta_i} is the vector of linear predictors associated with
ith row and  \mjeqn{\mathbb{E}\left(\frac{\partial^{2}\ell_{i}}{\partial\eta_{i}^{T}\partial\eta_{i}}\right)}{E(d^2l_i/deta_i^Tdeta_i)}
corresponds to weights associated with ith row and \mjeqn{\boldsymbol{W}}{W}
is a block matrix, made of diagonal matrixes \mjeqn{\mathbb{E}\left(\frac{\partial^{2}\ell}{\partial\boldsymbol{\eta}_{j}^{T}\partial\boldsymbol{\eta}_{i}}\right)}{E(d^2l/deta_j^Tdeta_i)}.
\item Regress \mjeqn{\boldsymbol{Z}}{Z} on \mjeqn{\boldsymbol{X}_{vlm}}{X_vlm}
to obtain \mjeqn{\boldsymbol{\beta}}{beta} as:
\mjdeqn{\boldsymbol{\beta}=\left(\boldsymbol{X}_{vlm}^{T}\boldsymbol{W}\boldsymbol{X}_{vlm}\right)^{-1}\boldsymbol{X}_{vlm}^{T}\boldsymbol{W}\boldsymbol{Z}}{(X_vlm^T W X_vlm)^-1(X_vlm^T W Z)}
\item Assign:
\itemize{
\item\code{converged <- }\mjeqn{\ell(\boldsymbol{\beta})-\ell_{-} < \ell_{-}\cdot\varepsilon}{l(beta)-l_-<l_- * epsilon} \code{or}
\mjeqn{||\boldsymbol{\beta}-\boldsymbol{\beta}_{-}||_{\infty} < \varepsilon}{||beta - beta_-||_inf< epsilon}
\item\code{iter <- iter + 1}
}
where \mjeqn{\varepsilon}{epsilon} is the relative tolerance level, by default \code{1e-8}.
\item Return to step 2.
\item Return \mjeqn{\boldsymbol{\beta}, \boldsymbol{W}}{beta, W}, \code{iter}.
}

In this package we use different conventions for \mjeqn{\boldsymbol{X}_vlm}{X_vlm}
matrix hence slight differences are present in algorithm description but
results are identical.
}
\references{
Yee, T. W. (2015). Vector Generalized Linear and Additive Models:
With an Implementation in R. New York, USA: Springer. ISBN 978-1-4939-2817-0.
}
\seealso{
\code{\link[stats:glm]{stats::glm()}} \code{\link[=estimate_popsize]{estimate_popsize()}} \code{\link[=control.method]{control.method()}} \code{\link[stats:optim]{stats::optim()}}
}
\author{
Piotr Chlebicki, Maciej Beręsewicz
}
