% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleRfit.R
\name{estimate_popsize.fit}
\alias{estimate_popsize.fit}
\title{Function to fit single-source capture-recapture models}
\usage{
estimate_popsize.fit(
  y,
  X,
  family,
  hwm,
  control,
  method,
  prior.weights,
  start,
  ...
)
}
\arguments{
\item{y}{vector of dependent variables}

\item{X}{model matrix}

\item{family}{same as model in estimate_popsize}

\item{hwm}{a vector containing information on how many columns in X matrix are associated with each linear predictor}

\item{control}{control parameters}

\item{method}{method of estimation same as in estimate_popsize}

\item{prior.weights}{vector of weights its the same argument as weights
in estimate_popsize}

\item{start}{start for regression fitting}

\item{...}{arguments to pass to other methods}
}
\value{
list of object connected to regression
}
\description{
Function to fit single-source capture-recapture models
}
