% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleRfit.R
\name{estimate_popsize.fit}
\alias{estimate_popsize.fit}
\title{Regression fitting for singleRcapture}
\usage{
estimate_popsize.fit(
  y,
  X,
  family,
  control,
  method,
  prior.weights,
  start,
  dispersion
)
}
\arguments{
\item{y}{vector of dependent variables}

\item{X}{model matrix}

\item{family}{same as model in estimate_popsize}

\item{control}{control parameters}

\item{method}{method of estimation same as in estimate_popsize}

\item{prior.weights}{vector of weights its the same argument as weights
in estimate_popsize}

\item{start}{start for regression fitting}

\item{dispersion}{start for dispersion}
}
\value{
list of object connected to regression
}
\description{
Regression fitting for singleRcapture
}
