% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleRmodel.R
\name{estimate_popsize}
\alias{estimate_popsize}
\title{estimate_popsize}
\usage{
estimate_popsize(
  formula,
  data,
  model = c("ztpoisson", "ztnegbin", "zotpoisson", "zotnegbin", "zelterman", "chao"),
  weights = 1,
  subset = NULL,
  na.action = NULL,
  trcount = 0,
  method = c("mle", "robust"),
  pop.var = c("analytic", "bootstrap"),
  control.method = NULL,
  control.model = NULL,
  control.pop.var = NULL,
  model.matrix = TRUE,
  x = FALSE,
  y = FALSE,
  contrasts = NULL,
  ...
)
}
\arguments{
\item{formula}{Description of model to that is supposed to be fitted}

\item{data}{Data frame for the regression}

\item{model}{Model for regression and population estimate}

\item{weights}{Optional object of a-priori weights used in fitting the model}

\item{subset}{As in lm/glm}

\item{na.action}{||}

\item{trcount}{Optional parameter for Zero-one truncated models, if population estimate
is for one inflated model then it specifies one counts and includes them in
final population estimate both point and interval, and for zeltermann/chao
estimator where it specifies counts of not used in estimate}

\item{method}{Method for fitting values currently supported IRLS and MaxLikelihood}

\item{pop.var}{A method of constructing confidence interval either analytic or bootstrap
where bootstraped confidence interval may either be based on 2.5\%-97.5\%
percientiles ("bootstrapPerc") or studentized CI ("bootstrapSD")}

\item{control.method}{||}

\item{control.model}{||}

\item{control.pop.var}{||}

\item{model.matrix}{If true returns model matrix as a part of returned object}

\item{x}{||}

\item{y}{||}

\item{contrasts}{||}

\item{...}{Arguments to be passed to other methods like subset in stats::model.frame}
}
\value{
Returns an object of classes inherited from glm containing:\cr

y argument used in regression\cr
formula provided on call\cr
Call on which model is based\cr
vector of fitt coefficients of regression\cr
vector of Standard errors of coeffcients\cr
vectors of test and p-values \cr
model used in regression\cr
akane information criterion\cr
vector of prior weights assingned on call\cr
vector of weights if IRLS was the fit method then its a vector what was returned by IRLS\cr
vector of working residuals\cr
Log-likelihood for the model\cr
number of itteration performed in fitting\cr
total and reduced degrees of freedom\cr
vector of fitt values\cr
a list containing population size estimate form PopulationEstimate function\cr
model, linear predictors, and qr decomposition of model matrix
}
\description{
estimate_popsize
}
