% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{farmsubmission}
\alias{farmsubmission}
\title{All farm submissions}
\format{
Data frame with 12036 rows and 4 columns.
\describe{
\item{TOTAL_SUB}{Number of submissions}
\item{log_size}{logarithm of size of farm}
\item{log_distance}{logarith of distance to nearest AHVLA center}
\item{C_TYPE}{Type of farm 1 for dairy and 0 for beef}
}
}
\usage{
farmsubmission
}
\description{
All farm submissions
}
\details{
Data on british animal farms submissions to AHVLA. It is required
for british farms to submit samples to AHVLA if unknown disease
if private veterinary surgeons cannot determine the cause of
death for the animal and decide to submit them, unless there is notifiable
disease suspected then it is required. This is data on such farms. This
data set contains farms that have submitted anything at all not only carcasses but
also blood samples etc.
}
\references{
Bohning et.al (2013) BIOMETRICS 69, 1033-1042
}
\keyword{datasets}
