% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{hatvalues.singleR}
\alias{hatvalues.singleR}
\title{Hat values for singleRclass}
\usage{
\method{hatvalues}{singleR}(model, ...)
}
\arguments{
\item{model}{object of clas singleRclass}

\item{...}{additional parameters to pass to other methods}
}
\value{
TODO
}
\description{
TODO
}
\details{
Hat values for singleRclass
}
