% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRLS.R
\name{IRLS}
\alias{IRLS}
\title{Iteratively Reweighted Least Squares}
\usage{
IRLS(
  Dependent,
  family,
  Covariates,
  start,
  disp = NULL,
  weights = NULL,
  Maxiter = 10000,
  eps = 1e-10
)
}
\arguments{
\item{Dependent}{A vector of Dependent variables}

\item{family}{Family of distributions used in regression}

\item{Covariates}{An array of Covariates}

\item{start}{A place to start the numerical estimation}

\item{disp}{dispersion parameter if family needs it}

\item{weights}{Optional object of weights used in fitting the model}

\item{Maxiter}{Maximum number of iterations}

\item{eps}{Precision level}
}
\value{
Returns a list of fitted Coefficients and number of iterations and final weights
}
\description{
A method for fitting regression parameters in generalized linear model
}
