% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRLS.R
\name{IRLS}
\alias{IRLS}
\title{Itterativly Reweighted Least Squares}
\usage{
IRLS(Dependent, Family, Covariates, start, Maxiter = 100, eps = 1e-10)
}
\arguments{
\item{Dependent}{A vector of Dependent variables}

\item{Family}{Family of distributions used in regression}

\item{Covariates}{An array of Covariates}

\item{start}{A place to start the numerical estimation}

\item{Maxiter}{Maximum number of itterations}

\item{eps}{Precision level}
}
\value{
A list of fitted Coefficients and number of itterations
}
\description{
A method for fitting regression parameters in generalised linear model
}
