% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal.R
\name{marginalFreq}
\alias{marginalFreq}
\title{Marginal Frequencies}
\usage{
marginalFreq(
  object,
  includeones = TRUE,
  includezeros = TRUE,
  onecount = NULL,
  range
)
}
\arguments{
\item{object}{Object of class singleR.}

\item{includeones}{Boolean value indicating whether to include the estimated number of zero counts.}

\item{includezeros}{Boolean value indicating whether to include one counts in the zero-one truncated models.}

\item{onecount}{A numeric value indicating number of one counts if null \code{trcount} from object will be assumed to be a number one counts.}

\item{range}{Optional argument specifying range of selected Y values.}
}
\value{
A table with marginal frequencies predicted by the model.
}
\description{
Marginal Frequencies
}
