% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal.R
\name{marginalFreq}
\alias{marginalFreq}
\title{Marginal Frequencies}
\usage{
marginalFreq(
  object,
  includeones = TRUE,
  includezeros = TRUE,
  onecount = NULL,
  range
)
}
\arguments{
\item{object}{object of class singleR}

\item{includeones}{Boolean value indicating whether to include estimated number of zero counts}

\item{includezeros}{Boolean value indicating whether to include one counts in zero one truncated models}

\item{onecount}{a numeric value indicating number of one counts if null trcount from object will be assumed to be a number one counts}

\item{range}{optional argument specifying range of selected Y values}
}
\value{
A table with marginal frequencies predicted by the model
}
\description{
Marginal Frequencies
}
