% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{netherlandsimmigrant}
\alias{netherlandsimmigrant}
\title{Netherlands immigrant data}
\format{
Data frame with 1880 rows and 5 columns.
\describe{
\item{capture}{Number of times a person has been captured}
\item{gender}{Gender for apprehended person 1 for male and 0 for female}
\item{age}{Age of apprehended person 0 for less than 40 yrs old 0 otherwise}
\item{reason}{Reason for being apprehended 1 if reason was illegally
residing in Netherlands 0 otherwise}
\item{nation}{Nation of origin of the captured person}
}
}
\usage{
data("netherlandsimmigrant")
}
\description{
Netherlands immigrant data
}
\details{
Data on irregular immigrants in Netherlands
}
\references{
van der Heijden et.al (2003) Statistical Modelling 3, 305-322
}
\keyword{datasets}
