% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstraps.R
\name{parBoot}
\alias{parBoot}
\title{Bootstraps for population size variation estimation}
\usage{
parBoot(
  family,
  y,
  X,
  dispersion,
  beta,
  weights,
  trcount,
  numboot,
  lambda,
  trace,
  control.bootstrap.method = NULL
)
}
\arguments{
\item{family}{TODO}

\item{y}{TODO}

\item{X}{TODO}

\item{dispersion}{TODO}

\item{beta}{TODO}

\item{weights}{TODO}

\item{trcount}{TODO}

\item{numboot}{TODO}

\item{lambda}{TODO}

\item{trace}{TODO}

\item{control.bootstrap.method}{TODO}
}
\value{
TODO
}
\description{
{
TODO
}
}
