% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.singleR}
\alias{plot.singleR}
\title{Diagnostic plots singleR regression}
\usage{
\method{plot}{singleR}(
  x,
  plotType = c("QQ", "marginal", "fitresid", "bootHist", "rootogram", "dfpopContr",
    "dfpopBox", "scaleLoc", "Cooks", "hatplot"),
  ...
)
}
\arguments{
\item{x}{object of singleRmargin class.}

\item{plotType}{Character parameter specifying type of plot to be made.
The following list presents and briefly explains possible type of plots:
\itemize{
\item \code{QQ} --
\item \code{marginal} --
\item \code{fitresid} --
\item \code{bootHist} --
\item \code{rootogram} --
\item \code{dfpopContr} --
\item \code{dfpopBox} --
\item \code{scaleLoc} --
\item \code{Cooks} --
\item \code{hatplot} --
}}

\item{...}{\itemize{
\item
\item
\item
\item
}}
}
\value{
TODO
}
\description{
Diagnostic plots singleR regression
}
