% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{popSizeEst}
\alias{popSizeEst}
\alias{popSizeEst.singleR}
\title{Title}
\usage{
popSizeEst(object, ...)

\method{popSizeEst}{singleR}(object, ...)
}
\arguments{
\item{object}{TODO}

\item{...}{TODO}
}
\value{
TODO
}
\description{
Title
}
