% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{popSizeEst}
\alias{popSizeEst}
\alias{popSizeEst.singleR}
\title{Extract population size estimation results}
\usage{
popSizeEst(object, ...)

\method{popSizeEst}{singleR}(object, ...)
}
\arguments{
\item{object}{Object with population size estimates}

\item{...}{TODO}
}
\value{
An object of class \code{popSizeEstResults} containing population size estimation results.
}
\description{
Extract population size estimation results
}
