% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{popSizeEst}
\alias{popSizeEst}
\alias{popSizeEst.singleR}
\title{Extract population size estimation results.}
\usage{
popSizeEst(object, ...)

\method{popSizeEst}{singleR}(object, ...)
}
\arguments{
\item{object}{object with population size estimates.}

\item{...}{additional optional arguments, currently not used in \code{singleR} class method.}
}
\value{
An object of class \code{popSizeEstResults} containing population size estimation results.
}
\description{
An extractor function with \code{singleR} method for extracting
important information regarding pop size estimate.
}
