% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smallMethods.R
\name{popSizeEst}
\alias{popSizeEst}
\alias{popSizeEst.singleRStaticCountData}
\title{Extract population size estimation results.}
\usage{
popSizeEst(object, ...)

\method{popSizeEst}{singleRStaticCountData}(object, ...)
}
\arguments{
\item{object}{object with population size estimates.}

\item{...}{additional optional arguments, currently not used in \code{singleRStaticCountData} class method.}
}
\value{
An object of class \code{popSizeEstResults} containing population size estimation results.
}
\description{
An extractor function with \code{singleRStaticCountData} method for extracting
important information regarding pop size estimate.
}
