% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationEstimate.R
\name{populationEstimate}
\alias{populationEstimate}
\title{Total Population Point and Interval estimate}
\usage{
populationEstimate(
  y,
  X,
  grad,
  parameter,
  beta,
  weights = 1,
  hessian,
  family,
  dispersion,
  method = "analytic",
  control
)
}
\arguments{
\item{y}{Observed values}

\item{X}{A matrix of covariates}

\item{grad}{A gradient of a model with respect to regression parameters}

\item{parameter}{An estimated Parameter for the model}

\item{beta}{Fitted regression values}

\item{weights}{If model is weighted weights of particular observations}

\item{hessian}{Hessian of a model}

\item{family}{Model Family}

\item{dispersion}{Estimated dispersion parameter
for truncated Negative binomial distributions}

\item{method}{A method of constructing confidence interval either analytic
to use formula for analytic CI or bootstrap where bootstraped confidence
interval may either be based on 2.5\%-97.5\% percientiles ("bootstrapPerc")
or by estimating SD ("bootstrapSD")}

\item{control}{list, same as control.pop.var in estimate_popsize}
}
\value{
Returns a list of size 3 with:
Point estimate, Interval estimate, and Variance
}
\description{
Creates a Horvitz-Thompson/Chao/Zelterman
point and interval estimate for total
and missing population from zero truncated poisson model.
}
