% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationEstimate.R
\name{PopulationEstimate}
\alias{PopulationEstimate}
\title{Total Population Point and Interval estimate}
\usage{
PopulationEstimate(
  y,
  X,
  Grad,
  Parameter,
  beta,
  weights = 1,
  Hess,
  family,
  Overdispersion = NULL,
  Method = "analytic"
)
}
\arguments{
\item{y}{Observed values}

\item{X}{A covariate matrix}

\item{Grad}{A gradient of a model with respect to regression parameters}

\item{Parameter}{An estimated Parameter for the model}

\item{beta}{Fitted regression values}

\item{weights}{If model is weighted weights of particular observations}

\item{Hess}{Hessian of a model}

\item{family}{Model Family}

\item{Overdispersion}{For truncated Negative binomial distributions}

\item{Method}{A method of constructing confidence interval either analytic or bootstrap where bootstraped confidence interval may
either be based on 2.5\%-97.5\% percientiles ("bootstrapPerc") or by estimating SD ("bootstrapSD") to use formula for analytic CI}
}
\value{
Returns a list of size 3 with: \cr
(1) Point estimate \cr
(2) Variance \cr
(3) Confidence interval constructed using Point estimate plus minus 1.96
(approximate of 97.5\% percentile of standardaised normal distribution) times standard deviation
}
\description{
Creates a Horwitz-Thompson point and interval estimate for total and missing population from zero truncated poisson model.
}
