% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationEstimate.R
\name{populationEstimate}
\alias{populationEstimate}
\title{Function to estimate population size based on single-source capture-recapture models}
\usage{
populationEstimate(
  y,
  X,
  grad,
  lambda,
  beta,
  weights = 1,
  weights0 = NULL,
  hessian,
  family,
  dispersion,
  method = "analytic",
  control
)
}
\arguments{
\item{y}{Observed values}

\item{X}{A matrix of covariates}

\item{grad}{A gradient of a model with respect to regression parameters}

\item{lambda}{An estimated lambda parameter for the model}

\item{beta}{Estimated parameters of the regression model (optional)}

\item{weights}{Case weights if applied}

\item{weights0}{Weights for all observations}

\item{hessian}{Hessian of a model}

\item{family}{Model family}

\item{dispersion}{Estimated dispersion parameter
for truncated Negative binomial distributions}

\item{method}{A method of constructing confidence interval either analytic
to use formula for analytic CI or bootstrap where bootstraped confidence
interval may either be based on 2.5\%-97.5\% percentiles or by bootstrap SE
estimation}

\item{control}{List argument, same as control.pop.var in estimate_popsize}
}
\value{
Returns a list of size 3 with:
Point estimate, Interval estimate, and Variance
}
\description{
Provides point and interval estimates for the population size based on variety of models
}
