% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationEstimate.R
\name{populationEstimate}
\alias{populationEstimate}
\title{Total Population Point and Interval estimate}
\usage{
populationEstimate(
  y,
  X,
  grad,
  parameter,
  beta,
  weights = 1,
  hessian,
  family,
  trcount,
  dispersion,
  method = "analytic"
)
}
\arguments{
\item{y}{Observed values}

\item{X}{A matrix of covariates}

\item{grad}{A gradient of a model with respect to regression parameters}

\item{parameter}{An estimated Parameter for the model}

\item{beta}{Fitted regression values}

\item{weights}{If model is weighted weights of particular observations}

\item{hessian}{Hessian of a model}

\item{family}{Model Family}

\item{trcount}{Optional parameter for Zero-one truncated models, if population estimate
is for one inflated model then it specifies one counts and includes them in
final population estimate both point and interval, and for zeltermann/chao
estimator where it specifies counts of not used in estimate}

\item{dispersion}{Estimated dispersion parameter
for truncated Negative binomial distributions}

\item{method}{A method of constructing confidence interval either analytic
to use formula for analytic CI or bootstrap where bootstraped confidence
interval may either be based on 2.5\%-97.5\% percientiles ("bootstrapPerc")
or by estimating SD ("bootstrapSD")}
}
\value{
Returns a list of size 3 with:
Point estimate, Interval estimate, and Variance
}
\description{
Creates a Horvitz-Thompson/Chao/Zelterman
point and interval estimate for total
and missing population from zero truncated poisson model.
}
