% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{predict.singleR}
\alias{predict.singleR}
\title{Predict method for \code{singleR} class}
\usage{
\method{predict}{singleR}(
  object,
  newdata,
  type = c("response", "link", "mean", "popSize", "contr"),
  se.fit = FALSE,
  na.action = NULL,
  weights,
  ...
)
}
\arguments{
\item{object}{a}

\item{newdata}{a}

\item{type}{a}

\item{se.fit}{a}

\item{na.action}{a}

\item{...}{a}
}
\value{
TODO
}
\description{
Predict method for \code{singleR} class
}
\seealso{
\code{\link[=redoPopEstimation]{redoPopEstimation()}} \code{\link[stats:summary.glm]{stats::summary.glm()}}
}
