% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{redoPopEstimation}
\alias{redoPopEstimation}
\alias{redoPopEstimation.singleR}
\title{Updating population size estimation results.}
\usage{
redoPopEstimation(object, ...)

\method{redoPopEstimation}{singleR}(object, cov = NULL, ...)
}
\arguments{
\item{object}{Object for which update of population size estimation results will be done.}

\item{...}{Additional optional arguments, currently not used in \code{singleR} class method.}

\item{cov}{An updated covariance matrix estimate.}
}
\value{
An object of class \code{popSizeEstResults} containing updated
population size estimation results.
}
\description{
A function that applies all post-hoc procedures that were taken
(such as heteroscedastic consistent covariance matrix estimation or bias
reduction) to population size estimation and standard error estimation.
}
\examples{
# Create simple model
Model <- estimatePopsize(
  formula = capture ~ nation + gender, 
  data = netherlandsimmigrant, 
  model = ztpoisson, 
  method = "IRLS"
)
# Apply heteroscedasticity consistent covariance matrix estimation
require(sandwich)
cov <- vcovHC(Model, type = "HC3")
summary(Model, cov = cov,
popSizeEst = redoPopEstimation(Model, cov = cov))
# Compare to results with usual covariance matrix estimation
summary(Model)
}
