% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{summary.singleR}
\alias{summary.singleR}
\title{summary.singleR}
\usage{
\method{summary}{singleR}(object, ...)
}
\arguments{
\item{object}{Object for which method is applied summarized}

\item{...}{other arguments to be passed to other methods}
}
\value{
Easy to read summary of regression and most important data from a singleR class
}
\description{
Method for class singleR, unlike glm and lm standard errors are
needed to estimate population size in main function,
so this only prints results and no object is returned.
}
\details{
Description of an object with class singleR, focusing on the associated
regression model and population size estimation performed
}
