% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{summary.singleR}
\alias{summary.singleR}
\title{summary.Model}
\usage{
\method{summary}{singleR}(object, ...)
}
\arguments{
\item{object}{Object for which method is aplied summarised}

\item{...}{other arguments to be passed to other methods}
}
\value{
Easy to read summary of regression and most important data from a 'SingleR' class
}
\description{
summary.Model
}
\details{
Describes an object of model class representing generalized linear regression
}
