% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{summary.singleRmargin}
\alias{summary.singleRmargin}
\title{Summary for marginal frequencies}
\usage{
\method{summary}{singleRmargin}(object, df = NULL, dropl5 = c("drop", "group", "no"), ...)
}
\arguments{
\item{object}{object of singleRmargin class}

\item{df}{degrees of freedom are sometimes not possible to automatically obtain if so this overwrites df of an object}

\item{dropl5}{boolean value indicating whether to group bins with frequencies < 5, drop them or do nothing}

\item{...}{Currently does nothing}
}
\value{
A chi squared test for comparison between fitted and observed marginal frequencies
}
\description{
Summary for marginal frequencies
}
