% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{vcov.singleR}
\alias{vcov.singleR}
\title{Obtain Covariance Matrix from Fitted singleR class Object}
\usage{
\method{vcov}{singleR}(object, type = c("Fisher", "observedInform"), ...)
}
\arguments{
\item{object}{object of singleR class.}

\item{type}{type of estimate for variance covariance matrix for now either
expected (Fisher) information matrix or observed information matrix.}

\item{...}{variables to pass to solve.}
}
\value{
A covariance matrix for fitted coefficients, rows and columns of which
correspond to parameters returned by \code{coef} method.
}
\description{
Returns a estimated covariance matrix for model coefficients
calculated from analytic hessian or fisher information matrix.
Covariance type is taken from control parameter that have been provided
on call that created object.
}
\details{
vcov method for singleR class

For now the covariance matrix is calculated
}
